+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-mastra
 
+## 1.3.14-alpha.0
+
+## 1.3.13
+
+## 1.3.13-alpha.3
+
 ## 1.3.13-alpha.2
 
 ## 1.3.13-alpha.1